2024/b01lers CTF 2024/web/pwnhub/requirements.txt
flask-unsign[wordlist]
jinja2
requests
